refactor: combine login/signup handler in single function
